requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
const-utils==0.0.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
numpy==2.1.1
packaging==24.1
propcache==0.2.1
PyYAML==6.0.2
regex==2024.9.11
safetensors==0.4.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
